requirements.txt
Pillow==6.2.2
requests==2.24.0
openslide-python==1.1.2
opencv-python==4.2.0.32
numpy==1.16.6
shapely==1.7.1
